requirements.txt
# Cleaned requirements file - conda packages removed
# Original file: requirements_controlvla.txt

absl-py==2.1.0
addict==2.4.0
aiohttp-cors==0.7.0
annotated-types==0.7.0
anyio==4.8.0
asciitree==0.3.3
asyncio==3.4.3
atomics==1.0.2
blessed==1.20.0
blinker==1.9.0
cachetools==5.3.3
# camera-calibration==1.17.0
# camera-calibration-parsers==1.12.0
# catkin==0.8.10
# catkin-pkg==1.0.0
chardet==5.2.0
colorful==0.5.6
colorlog==6.9.0
ConfigArgParse==1.7
# controller-manager==0.20.0
# controller-manager-msgs==0.20.0
# cv-bridge==1.16.2
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==4.4.2
# diagnostic-analysis==1.11.0
# diagnostic-common-diagnostics==1.11.0
# diagnostic-updater==1.11.0
# diff_gaussian_rasterization @ git+https://github.com/dcharatan/diff-gaussian-rasterization-modified@1250c420ebb945f0dce9945086e22faab9157c92
diffusers==0.32.2
distlib==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
docutils==0.21.2
# dynamic-reconfigure==1.7.3
easydict==1.13
egl_probe==1.0.2
embreex==2.17.7.post6
etils==1.5.2
eval_type_backport==0.2.2
evdev==1.7.1
fastapi==0.115.11
fastcrc==0.2.0
fasteners==0.15
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
ffmpy==0.5.0
Flask==3.0.3
# free-mujoco-py==2.1.6
freetype-py==2.4.0
future==1.0.0
# gazebo_plugins==2.9.2
# gazebo_ros==2.9.2
# gencpp==0.7.0
# geneus==3.0.0
# genlisp==0.4.18
# genmsg==0.6.0
# gennodejs==2.0.2
# genpy==0.6.15
gevent==24.2.1
glfw
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
gpustat==1.1.1
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
hidapi==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hydra-core==1.3.2
# image-geometry==1.16.2
imagecodecs==2023.9.18
imageio==2.34.1
imageio-ffmpeg==0.5.0
# interactive-markers==1.12.0
iopath==0.1.10
ipdb==0.13.13
itsdangerous==2.2.0
# joint-state-publisher==1.15.1
# joint-state-publisher-gui==1.15.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
# laser_geometry==1.6.7
lightning==2.5.0.post0
lightning-utilities==0.12.0
llvmlite==0.42.0
loguru==0.7.2
lpips==0.1.4
lxml==5.3.0
manifold3d==3.0.1
mapbox_earcut==1.0.3
Markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
# message-filters==1.16.0
monotonic==1.6
moviepy==1.0.3
msgspec==0.19.0
mujoco==3.1.0
mujoco_ar==1.3.0
munkres==1.1.4
narwhals==1.29.0
nbformat==5.10.4
networkx
nodeenv==1.9.1
numba==0.59.1
numpy==1.26.4
nvidia-ml-py==12.535.161
nvitop==1.3.2
open3d==0.19.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-contrib-python==4.11.0.86
opencv-python==4.7.0.72
orjson==3.10.15
paramz==0.9.5
pathtools==0.1.2
platformdirs==3.11.0
plotly==6.0.0
plyfile==1.1
portalocker==2.10.1
proglog==0.1.10
prometheus_client==0.20.0
proto-plus==1.23.0
protobuf==4.23.3
py-spy==0.3.14
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybullet==3.2.6
pycollada==0.6
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
PyExifTool==0.5.0
pyglet==2.0.17
pynput==1.7.6
PyOpenGL==3.1.0
PyQt5==5.15.9
PyQt5-sip==12.12.2
pyquaternion==0.9.9
pyrealsense2==2.55.1.6486
pyrender==0.1.45
python-multipart==0.0.20
# python-qt-binding==0.4.4
python-xlib==0.33
pytorch-lightning==2.5.0.post0
# qt-dotgraph==0.4.2
# qt-gui==0.4.2
# qt-gui-cpp==0.4.2
# qt-gui-py-common==0.4.2
ray==2.8.0
referencing==0.35.1
# resource_retriever==1.12.7
retrying==1.3.4
rich==13.9.4
robomimic==0.2.0
robosuite @ https://github.com/cheng-chi/robosuite/archive/3f2c3116f1c0a1916e94bfa28da4d055927d1ab3.tar.gz
# rosbag==1.16.0
# rosboost-cfg==1.15.8
# rosclean==1.15.8
# roscreate==1.15.8
# rosgraph==1.16.0
# roslaunch==1.16.0
# roslib==1.15.8
# roslint==0.12.0
# roslz4==1.16.0
# rosmake==1.15.8
# rosmaster==1.16.0
# rosmsg==1.16.0
# rosnode==1.16.0
# rosparam==1.16.0
# rospkg==1.5.1
# rospy==1.16.0
# rosservice==1.16.0
# rostest==1.16.0
# rostopic==1.16.0
# rosunit==1.15.8
# roswtf==1.16.0
rpds-py==0.18.1
# rqt-console==0.4.12
# rqt-image-view==0.4.17
# rqt-logger-level==0.4.12
# rqt-moveit==0.5.11
# rqt-reconfigure==0.5.5
# rqt-robot-dashboard==0.5.8
# rqt-robot-monitor==0.5.15
# rqt-runtime-monitor==0.5.10
# rqt-rviz==0.7.0
# rqt-tf-tree==0.6.4
# rqt_action==0.4.9
# rqt_bag==0.5.1
# rqt_bag_plugins==0.5.1
# rqt_dep==0.4.12
# rqt_graph==0.4.14
# rqt_gui==0.5.3
# rqt_gui_py==0.5.3
# rqt_launch==0.4.9
# rqt_msg==0.4.10
# rqt_nav_view==0.5.7
# rqt_plot==0.4.13
# rqt_pose_view==0.5.11
# rqt_publisher==0.4.10
# rqt_py_common==0.5.3
# rqt_py_console==0.4.10
# rqt_robot_steering==0.5.12
# rqt_service_caller==0.4.10
# rqt_shell==0.4.11
# rqt_srv==0.4.9
# rqt_top==0.4.10
# rqt_topic==0.4.13
# rqt_web==0.4.10
rsa==4.9
Rtree==1.3.0
ruff==0.9.9
# rviz==1.14.25
scikit-image==0.24.0
scikit-video==1.1.11
semantic-version==2.10.0
# sensor-msgs==1.13.1
shapely==2.0.7
shellingham==1.5.4
shtab==1.7.1
simplejson==3.19.2
# smach==2.5.2
# smach-ros==2.5.2
smart-open==7.0.4
# smclib==1.8.6
sniffio==1.3.1
starlette==0.46.0
svg.path==6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
# tf==1.13.2
# tf-conversions==1.13.2
# tf2-geometry-msgs==0.7.7
# tf2-kdl==0.7.7
# tf2-py==0.7.7
# tf2-ros==0.7.7
threadpoolctl==3.2.0
tifffile==2024.8.30
tokenizers==0.15.2
tomlkit==0.12.0
# topic-tools==1.16.0
torchmetrics==1.6.1
tqdm==4.66.5
transformers==4.38.2
trimesh==4.4.6
triton==2.1.0
typeguard==4.4.2
typer==0.15.2
typing_extensions==4.12.2
tyro==0.9.16
# ur-rtde==1.5.6
# urdfpy==0.0.22
uritemplate==4.1.1
uvicorn==0.34.0
v4l2py==2.2.0
vhacdx==0.0.8.post2
virtualenv==20.21.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
# xacro==1.14.18
xatlas==0.0.9
xxhash==3.5.0
yourdfpy==0.0.57
zerorpc==0.6.3
zope.event==5.0
zope.interface==6.4.post2

# Converted from conda_environment.yaml to pip requirements format
# Core scientific computing
numpy
scipy==1.11
numba==0.57
pandas==2.1

# Development and visualization
ipykernel==6.23
matplotlib==3.7
k3d==2.15
pyyaml==6.0
tqdm==4.65

# Data handling and storage
zarr==2.16
numcodecs==0.11

# Machine learning and AI
hydra-core==1.2.0
wandb==0.15.8
dill==0.3.7
einops==0.6.1
diffusers==0.18.2
timm==0.9.7
accelerate==0.24

# Video and media processing
av==10.0
scikit-video==1.1.11

# Scientific computing
scikit-fda==0.9

# Cloud and web services
boto3==1.33
aiohttp==3.9
aiofiles==23.2

huggingface-hub==0.23.2